Refactor rsvp page
