+++ source/scenes/Rsvp/Panes/LanguageInfo.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .language-info {
   display: flex;
