+++ source/scenes/Rsvp/Panes/SubmitStyles.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .submission-page {
   .submit-button,
