+++ source/scenes/Rsvp/Panes/TechInterests.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .ti-inst {
   text-align: center;
