--- source/scenes/Rsvp/components/Form/styles.scss
@@ -1,154 +1,0 @@
-@import '../../../../styles/variables.scss';
-
-$form-width: 500px;
-
-.scrollable-form {
-  background-image: linear-gradient(
-    0deg,
-    #017680 0%,
-    #1d858e 16%,
-    #1f868f 18%,
-    #349199 49%,
-    #2a8c94 54%,
-    #0c7c86 72%,
-    #017680 82%
-  );
-  flex-grow: 1;
-  overflow: hidden;
-  position: relative;
-
-  h1 {
-    position: relative;
-    padding-top: 62px;
-    color: white;
-    text-align: center;
-    font-weight: 700;
-    font-size: 44px;
-    z-index: 5;
-  }
-
-  #form-view {
-    margin: auto;
-    padding: 60px 40px 0 40px;
-    position: relative;
-    z-index: 5;
-    max-width: $form-width;
-
-    .form-transition {
-      position: relative;
-      opacity: 0;
-    }
-
-    .hidden {
-      display: none;
-    }
-
-    .split {
-      display: flex;
-      flex-direction: row;
-
-      > *:first-child {
-        margin-right: 20px;
-      }
-
-      > *:last-child {
-        margin-left: 20px;
-      }
-    }
-  }
-
-  .nav-buttons {
-    display: flex;
-    flex-direction: row;
-    margin: auto;
-    opacity: 0;
-    padding-top: 20px;
-    position: absolute;
-    top: 100px;
-    width: calc(100% - 80px);
-
-    button {
-      border: 2px solid white;
-      border-radius: 100px;
-      color: white;
-      cursor: pointer;
-      font-size: 16px;
-      font-weight: bold;
-      padding: 6px 18px;
-      transition: color 0.2s, border-color 0.2s;
-
-      &:disabled {
-        border-color: $white-greyed;
-        color: $white-greyed;
-        cursor: unset;
-      }
-    }
-
-    .spacing {
-      flex-grow: 1;
-    }
-  }
-
-  .rocks {
-    position: absolute;
-    top: -2vh;
-    z-index: 2;
-    height: 104vh;
-  }
-
-  .left-rocks {
-    left: -1px;
-  }
-  .right-rocks {
-    right: -1px;
-  }
-  .currents {
-    position: absolute;
-    top: 0;
-    left: 0;
-    z-index: 0;
-    width: 100%;
-    min-width: 1100px;
-    opacity: 0.2;
-  }
-  .fish1 {
-    position: absolute;
-    bottom: 60px;
-    right: 45px;
-    width: 150px;
-  }
-  .fish2 {
-    position: absolute;
-    top: 30px;
-    left: 0;
-    width: 250px;
-  }
-
-  @media (max-height: 760px) {
-    h1 {
-      padding-top: 50px;
-    }
-
-    #form-view {
-      padding-top: 15px;
-    }
-  }
-
-  @media (max-height: 680px) {
-    h1 {
-      padding-top: 35px;
-    }
-
-    .form-field {
-      padding-bottom: 10px;
-
-      input {
-        padding-top: 0px;
-      }
-    }
-  }
-
-  .right-rocks {
-    right: -12px;
-  }
-}
